Update for v2
All I can tell from this is that it's now very complicated to run lots
of package resources!

I think it may be worth implementing some sort of Async field or Lock
function to a resource. We could create a new named lock as a global var:

```
var AptLock = foo.NewLock()
```

When we create resources we say "try to run, but only if there is a lock
available":

```
foo.WithLock(foo.Create(v.P("cowsay")), AptLock)
```